Add utility modules for dlist and VLAN bitmaps
